backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-dotenv==1.0.1
openai==1.12.0
faster-whisper==1.0.0
edge-tts==6.1.10
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.4
pydantic==2.6.1
